Move BigQueryNG to types
